feat(ssa): add ssa module with pass manager
