+++ src/ssa/error.rs
@@ -2,12 +2,22 @@ use thiserror::Error;
 
 #[derive(Debug, Error)]
 pub enum SsaError {
+    #[error("function has already been built")]
+    FunctionAlreadyBuilt,
+    #[error("referenced function not found")]
+    FunctionNotFound,
+    #[error("incorrect number of arguments received for function")]
+    IncorrectFunctionArgs,
+    #[error("argument of invalid type received for function")]
+    IncorrectFuncArgType,
     #[error("referenced instruction not found")]
     InstrNotFound,
     #[error("encountered an unexpectedly filled block")]
     BlockFilled,
     #[error("block was not filled when it was expected to be")]
     BlockNotFilled,
+    #[error("block was not sealed when it was expected to be")]
+    BlockNotSealed,
     #[error("block was not empty when it was expected to be")]
     BlockNotEmpty,
     #[error("referenced block not found")]
